backend/requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
Flask-SocketIO==5.5.1
python-socketio==5.14.1
python-dotenv==1.1.1
spotipy==2.25.1
mysql-connector-python==9.4.0
opencv-python==4.12.0.88

# Required by mood detection
numpy
pillow
mediapipe



